### 2019-01-21 #1190 @whedon
[REVIEW]: Multiblock PLS: Block dependent prediction modeling for Python
**Submitting author:** @lvermue (<a href="http://orcid.org/0000-0002-3403-7624">Laurent Vermue</a>)
**Repository:** <a href="https://github.com/DTUComputeStatisticsAndDataAnalysis/MBPLS" target ="_blank">https://github.com/DTUComputeStatisticsAndDataAnalysis/MBPLS</a>
**Version:** v1.0.0
