### 2019-01-22 #1190 @arokem
@lvermue: first of all, kudos on a nice and useful software package!

Could you please share the code that you used to generate the results in the "Benchmark" section of your paper? Thanks!
