### 2019-01-22 #1190 @arokem
I have looked over all of the review criteria (also shown above) and posted issues to your repository marked with "[JOSS review]". I also added some other issues, that are not required for the review of your paper, but I still think would be good additions or changes 😄
