### 2019-01-29 1cc8ac6
Create bug report and feature request templates
