+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,34 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+[A clear and concise description of what the bug is.]
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. [First step]
+2. [Second step]
+3. [...]
+4. [See error]
+
+**Expected behavior**
+[A clear and concise description of what you expected to happen.]
+
+**Environment (please complete the following information):**
+ - OS: [e.g. macOS]
+ - Python version: [e.g. 3.6.5]
+ - Version of this software: [e.g. 1.0.0]
+ - Versions of required Python packages:
+   - Numpy: [e.g. 1.13.3]
+   - Scipy: [e.g. 1.0.0]
+   - Scikit-learn: [e.g. 0.20.0]
+   - Pandas: [e.g. 0.20.0]
+
+**Additional information**
+[Add any other information about the problem here.]
