### 2019-01-29 923a500
Create pull request templates 	new file:   .github/PULL_REQUEST_TEMPLATE/bug_fix.md 	new file:   .github/PULL_REQUEST_TEMPLATE/feature_change.md 	new file:   .github/pull_request_template.md
