+++ .github/pull_request_template.md
@@ -1,0 +1,8 @@
+Welcome contributor 👋. Please follow these steps to submit a pull request.
+
+1. Copy the right template
+   - For fixing a bug: [Bug fix template](https://raw.githubusercontent.com/DTUComputeStatisticsAndDataAnalysis/MBPLS/master/.github/PULL_REQUEST_TEMPLATE/bug_fix.md)
+   - For new features: [Feature request template](https://raw.githubusercontent.com/DTUComputeStatisticsAndDataAnalysis/MBPLS/master/.github/PULL_REQUEST_TEMPLATE/feature_change.md)
+2. Replace this text with the new template
+3. Fill in all requested information from the template
+4. Submit the pull request 🎉
