+++ .github/PULL_REQUEST_TEMPLATE/feature_change.md
@@ -1,0 +1,54 @@
+## Contributing to feature requests
+
+ - Fill out the template below
+ - The pull request should refer to an existing feature request that has the label 'help wanted' or 'accepted'
+ - The pull request should comply with these points mentioned in [the contribution guidelines](https://github.com/DTUComputeStatisticsAndDataAnalysis/MBPLS/blob/master/.github/CONTRIBUTING.md#pull-request)
+ 
+ ### Link to the relevant feature request
+ <!--
+
+Link to the issue describing the feature request. The affected feature request should have the label 'help wanted' or
+'accepted'.
+
+-->
+ 
+### Description of the Change
+
+<!--
+
+We must be able to understand the design of your change from this description, so please walk us through the concepts.
+
+-->
+
+### Alternatives
+
+<!-- Explain what other solutions you have considered to realize this functionality.  -->
+
+### Possible Drawbacks
+
+<!-- What are the possible side-effects or negative impacts of the code change? -->
+
+### Verification Process
+
+<!--
+
+What process did you follow to verify that the change has not introduced any regressions? Describe the actions you
+performed and describe the results you observed.
+
+-->
+
+### Release Notes
+
+<!--
+
+Please describe the changes in a single line that explains this improvement in
+terms that a user can understand.
+
+Example:
+
+- Fixed an issue where model parameters were not set correctly.
+- Increased the performance of method X by avoiding redundant re-calculations of the loading vectors.
+
+-->
+ 
+
