+++ .github/PULL_REQUEST_TEMPLATE/bug_fix.md
@@ -1,0 +1,52 @@
+## Contributing to bug fixes
+
+ - Fill out the template below
+ - The pull request should only fix existing bug reports
+ - The pull request should comply with these points mentioned in [the contribution guidelines](https://github.com/DTUComputeStatisticsAndDataAnalysis/MBPLS/blob/master/.github/CONTRIBUTING.md#pull-request)
+ 
+ ### Link to the relevant Bug(s)
+ <!--
+
+Link to the issue describing the bug that you're fixing.
+
+If there is not yet an issue for your bug, please open a new issue and then link to that issue in your pull request.
+Note: In some cases, one person's "bug" is another person's "feature." If the pull request does not address an existing
+issue with the "bug" label, the maintainers have the final say on whether the current behavior is a bug.
+
+-->
+ 
+### Description of the Change
+
+<!--
+
+We must be able to understand the design of your change from this description, so please walk us through the concepts.
+
+-->
+
+### Possible Drawbacks
+
+<!-- What are the possible side-effects or negative impacts of the code change? -->
+
+### Verification Process
+
+<!--
+
+What process did you follow to verify that the change has not introduced any regressions? Describe the actions you
+performed and describe the results you observed.
+
+-->
+
+### Release Notes
+
+<!--
+
+Please describe the changes in a single line that explains this improvement in
+terms that a user can understand.
+
+Example:
+
+- Fixed an issue where model parameters were not set correctly.
+- Increased the performance of method X by avoiding redundant re-calculations of the loading vectors.
+
+-->
+ 
