### 2019-01-29 9181d72
Create contribution guidelines and Code of Conduct 	new file:   .github/CODE_OF_CONDUCT.md 	new file:   .github/CONTRIBUTING.md
