+++ .github/CONTRIBUTING.md
@@ -1,0 +1,61 @@
+# Contributing
+
+Welcome supporter of this project! 👋 We are happy that you found your way to this page and appreciate every
+contribution, helping us to develop new features or fixing potential issues of this software.
+In order to keep this process transparent and allow us to keep track of things, we would like you to follow the
+following guidelines:
+
+1. [Report issues and problems or request new new features](#issues-and-feature-requests)
+2. [Contribute to the software](#pull-request)
+3. [Seek support or asking a question](#support)
+
+
+## Issues and Feature Requests
+To report errors in the software follow the steps for [bug reports](#bug-report). In case you would like to
+request a useful feature or add-on for this software you should follow the steps for
+[feature requests](#feature-request). 
+
+### Bug report
+We track bugs for this software as [GitHub Issues](https://guides.github.com/features/issues/).
+To file a bug report, please [submit an issue](#https://github.com/DTUComputeStatisticsAndDataAnalysis/MBPLS/issues/new?assignees=&labels=&template=bug_report.md&title=)
+and provide the therein requested information.
+
+In case you don't have a GitHub account, you can use [this template](https://github.com/DTUComputeStatisticsAndDataAnalysis/MBPLS/blob/master/.github/ISSUE_TEMPLATE/bug_report.md)
+and send it via e-mail to either [andba@dtu.dk](mailto:andba@dtu.dk) or [lauve@dtu.dk](mailto:lauve@dtu.dk).
+
+    
+### Feature request
+We track feature requests for this software as [GitHub Issues](https://guides.github.com/features/issues/).
+To file a feature request, please [submit an issue](#https://github.com/DTUComputeStatisticsAndDataAnalysis/MBPLS/issues/new?assignees=&labels=&template=feature_request.md&title=)
+and provide the therein requested information.
+
+In case you don't have a GitHub account, you can use [this template](https://github.com/DTUComputeStatisticsAndDataAnalysis/MBPLS/blob/master/.github/ISSUE_TEMPLATE/feature_request.md)
+and send it via e-mail to either [andba@dtu.dk](mailto:andba@dtu.dk) or [lauve@dtu.dk](mailto:lauve@dtu.dk).
+
+
+## Pull request
+Requests for direct code changes are handled as 
+[GitHub Pull Requests](https://help.github.com/articles/about-pull-requests/) and can be
+[submitted here](https://github.com/DTUComputeStatisticsAndDataAnalysis/MBPLS/compare).
+
+Before submitting pull requests, please make sure that these address existing bug reports and/or feature requests. If this
+is not the case, please make sure to first create these as described in
+[Issues and Feature Requests](#issues-and-feature-requests).
+
+When submitting pull requests, we encourage you to comply with these points:
+ - Follow all instructions within the [pull request template](https://github.com/DTUComputeStatisticsAndDataAnalysis/MBPLS/blob/master/.github/pull_request_template.md)
+ - Make sure that you create or adjust all documentation that is affected by your code changes
+ - Create test scripts for non-trivial code changes, if these are not covered by existing tests
+
+
+    
+## Support
+In case you just want to ask a question or need support for this software, you can reach out to the project team 
+via e-mail to either [andba@dtu.dk](mailto:andba@dtu.dk) or [lauve@dtu.dk](mailto:lauve@dtu.dk).
+
+## Code of Conduct
+
+This project and everyone contributing to it is governed by
+[our Code of Conduct](https://github.com/DTUComputeStatisticsAndDataAnalysis/MBPLS/blob/master/.github/CODE_OF_CONDUCT.md).
+By participating, you are expected to comply with this code. Please report unacceptable behavior to the project team at
+[andba@dtu.dk](mailto:andba@dtu.dk) or [lauve@dtu.dk](mailto:lauve@dtu.dk).
