### 2019-01-29 caa8c31
Rename the method testing script
Fixes #5
	modified:   mbpls/tests/__init__.py
	renamed:    mbpls/tests/test_Installation.py -> mbpls/tests/test_mbpls.py
