### 2019-01-29 cf5e5d8
Add DOIs to references, where available
Fixes #4
	modified:   paper/paper.bib
