+++ paper/paper.bib
@@ -5,7 +5,8 @@
     year = {1998},
     volume = {12},
     number = {5},
-    pages = {301--321}
+    pages = {301--321},
+    doi = {10.1002/(SICI)1099-128X(199809/10)12:5<301::AID-CEM515>3.0.CO;2-S},
 }
 
 @Article{rannar:95,
@@ -16,6 +17,7 @@
     volume={9},
     number={6},
     pages={459--470},
+    doi = {10.1002/cem.1180090604},
 }
 
 @Article{rannar:94,
@@ -26,6 +28,7 @@
     volume={8},
     number={2},
     pages={111--125},
+    doi = {10.1002/cem.1180080204},
 }
 
 @Article{li:16,
@@ -36,6 +39,7 @@
     volume={19},
     number={2},
     pages={325--340},
+    doi = {10.1093/bib/bbw113},
 }
 
 @Book{martens:01,
@@ -54,6 +58,7 @@
     volume={2},
     number={6},
     pages={349--376},
+    doi = {10.1016/0898-5529(89)90004-3},
 }
 
 @Article{pedregosa:11,
@@ -74,6 +79,7 @@
     volume={18},
     number={3},
     pages={251--263},
+    doi = {10.1016/0169-7439(93)85002-X},
 }
 
 @Article{bougeard:18,
@@ -84,6 +90,7 @@
     volume = {86},
     number = {1},
     pages={1--17},
+    doi = {10.18637/jss.v086.i01},
 }
 
 @Article{wold:84,
@@ -94,6 +101,7 @@
     volume={5},
     number={3},
     pages={735--743},
+    doi = {10.1137/0905052},
 }
 
 @Article{lindgren:93,
@@ -104,4 +112,5 @@
     volume={7},
     number={1},
     pages={45--59},
+    doi = {10.1002/cem.1180070104},
 }
