### 2019-01-29 8fc338d
Rearrange documentation and example link position in the README
Fixes #3
	modified:   README.rst
