### 2019-01-31 #1190 @lvermue
@arokem Thank you for your comments and useful suggestions!🙂
We have worked on all issues and feel confident that we have resolved them.

Regarding your previous comment:
> Could you please share the code that you used to generate the results in the "Benchmark" section of your paper? Thanks!

The zip-file attached to this comment contains three different python scripts, e.g. for testing run-times for fixed row size with increasing column size, fixed column size with increasing row size as well as a symmetric increase of both row and column sizes as reported in the paper.

The scripts also contain the tests for the Ade4-package, which was run through rpy2. We also performed various tests with pure R scripts, but could not detect any difference in performance nor in the times recorded.

The redis redlock module was used to let 4 identical servers coordinate and work on the test script simultaneously, i.e. each server knows what the others have done and are doing at the moment, so the server can pick a remaining task not handled by other servers.🤓
[runtime_analysis.zip](https://github.com/openjournals/joss-reviews/files/2819124/runtime_analysis.zip)
