### 2019-02-01 #1190 @arokem
Hi @lvermue : nice work on these revisions. Most of my comments are addressed.

Regarding the performance benchmark: any reason not to add these to the repo of your software? Or, barring that, put that on another publicly available software that you could refer to in your manuscript? That would improve the reproducibility of these benchmark results (I am already running into some issues running the scripts...)
