### 2019-02-03 13007ac
Add runtime analysis scripts used for the benchmark section of the paper
The scripts used to obtain the results shown in the benchmark section of
the paper were added as well as a brief README file showing the required
setup to run the tests.
	new file:   paper/benchmark/README.md
	new file:   paper/benchmark/runtime_analysis_colincrease.py
	new file:   paper/benchmark/runtime_analysis_rowincrease.py
	new file:   paper/benchmark/runtime_analysis_symmetric.py
