+++ paper/benchmark/runtime_analysis_colincrease.py
@@ -1,0 +1,237 @@
+from mbpls.mbpls import MBPLS
+import time
+import rpy2
+
+# define Parameters
+rand_seed = 25
+
+# initialize
+from matplotlib import pyplot as plt
+import numpy as np
+import pickle
+import pandas as pd
+
+np.seterr(all='raise')
+
+
+def mbpls_ade4(data, num_comp=20):
+    from rpy2 import robjects
+    from rpy2.robjects import pandas2ri
+    import time
+
+    x1, x2, y = data
+
+    pandas2ri.activate()  # to activate easy conversion from r to pandas dataframes
+
+    # to generate a dataframe in R global environment
+    robjects.globalenv['x1'] = pandas2ri.py2ri_pandasdataframe(x1)
+    robjects.globalenv['x2'] = pandas2ri.py2ri_pandasdataframe(x2)
+    robjects.globalenv['ref'] = pandas2ri.py2ri_pandasdataframe(y)
+
+    start_time = time.time()
+    # how to execute R code using the global environment variables defined above
+    robjects.r(
+        '''
+        library(ade4)
+        library(adegraphics)
+
+        dudiY.act <- dudi.pca(ref, center = TRUE, scale = TRUE, scannf =
+                                  FALSE,nf=1)
+        ktabX.act <- ktab.list.df(list(mainPerformance=x1,mainInput=x2))
+        resmbpls.act <- mbpls(dudiY.act, ktabX.act, scale = TRUE,
+                                option = "none", scannf = FALSE, nf=''' + str(num_comp) + ''')
+
+    bip <- resmbpls.act$bip
+    ''')
+
+    return start_time
+
+###################################################################
+# General settings
+###################################################################
+
+# How many runs?
+reruns = 3
+
+number_components = 20
+
+feature_size_y = 10
+
+standardize = True
+
+# Should all parameters be calculated
+calc_all = False
+
+# Verbosity
+# 0-show nothing
+# 1-show datasets
+# 2-show methods
+# 3-show run information
+# 4-show time of each run
+verbosity_level = 4 #higher verbosity levels includer lower ones
+
+
+# Calculation Measurements
+include_timing = True
+
+# Datasets
+x_sizes = np.arange(1000, 1001, 1000)
+y_sizes = np.arange(100, 20001, 100)
+
+
+methods = [
+    ('Ade4', 'Ade4'),
+    ('Unipals', 'UNIPALS'),
+    ('Kernel', 'KERNEL'),
+    ('Nipals', 'NIPALS'),
+    ('Simpls', 'SIMPLS')
+]
+
+
+###################################################################
+# General settings - END
+###################################################################
+class bcolors:
+    HEADER = '\033[95m'
+    OKBLUE = '\033[94m'
+    OKGREEN = '\033[92m'
+    WARNING = '\033[93m'
+    FAIL = '\033[91m'
+    ENDC = '\033[0m'
+    BOLD = '\033[1m'
+    UNDERLINE = '\033[4m'
+
+# Dictionary to save timing
+if include_timing:
+    timing_results = np.zeros(shape=(len(methods), len(y_sizes), len(x_sizes), reruns))
+    indicator_matrix = np.zeros(shape=(len(y_sizes), len(x_sizes)))
+
+initiate_files = True
+if initiate_files:
+    pickle.dump(indicator_matrix, open('indicator_matrix_colincrease.pkl', 'wb'))
+    pickle.dump(timing_results, open('timing_results_colincrease.pkl', 'wb'))
+
+# Derive shapes from datasets
+datasets_shape = []
+
+finished = False
+
+# load the indicator matrix that shows what has been done already
+# Lock the access first
+while not finished:
+    while True:
+        print('Loading indicator matrix data')
[…]
