+++ paper/benchmark/README.md
@@ -1,0 +1,11 @@
+To run the tests shown in the Benchmark section of the paper, the following additional requirements have to be
+fulfilled:
+1. Install R 
+2. Install the following R-packages
+   1. 'ade4'
+   2. 'sp'
+   3. 'adegraphics'
+3. Install rpy2
+4. Set the path to the R installation
+   - import os
+   - os.environ['R_HOME'] = '/Your/path/to/lib/R'

[…]
