### 2019-02-03 #1190 @lvermue
The initial reason not to add the runtime analysis scripts to our repo was that they were not considered as a part of the software itself. However, as they are a part of the paper I agree with you and have now added them as a subdirectory called 'benchmark' within the paper directory, including a small README showing the requirements to run the tests. Furthermore, the scripts were cleaned up and changed to run on one node only, not requiring the Redis lock setup for distributed calculations.
See https://github.com/DTUComputeStatisticsAndDataAnalysis/MBPLS/commit/13007ac8d19612ac06731b95c686fc56b916d7c4
