### 2019-02-06 #1190 @arokem
@brainstorm: I am ready to check that last box, and from my point of view this paper is ready to be accepted for publication.

@lvermue : Thanks for adding the code. I think that it benefits the paper greatly.

One small thing: I don't think that this generates the plots that appear in the paper. I realize that this is just one small additional step, but I think that it would be helpful for readers to see how you would get from your benchmark code to these plots. But take that as a recommendation, rather than as a requirement for my review of the paper.
