### 2019-02-06 abc9bf5
Add code to generate plots as shown in the paper
The runtime analysis scripts now contain the code used to generate the
plots shown within the paper.
	modified:   paper/benchmark/runtime_analysis_colincrease.py
	modified:   paper/benchmark/runtime_analysis_rowincrease.py
	modified:   paper/benchmark/runtime_analysis_symmetric.py
