+++ paper/benchmark/runtime_analysis_symmetric.py
@@ -72,6 +72,9 @@ verbosity_level = 4 #higher verbosity levels includer lower ones
 # Calculation Measurements
 include_timing = True
 
+# Plot timing graphs in the end
+plot_results = True
+
 # Datasets
 x_sizes = np.arange(100, 20001, 100)
 y_sizes = np.arange(100, 20001, 100)
@@ -234,3 +237,36 @@ while not finished:
     pickle.dump(indicator_matrix_shared, open('indicator_matrix_symmetric.pkl', 'wb'))
     pickle.dump(timing_results_shared, open('timing_results_symmetric.pkl', 'wb'))
 
+# Generate plot as shown in Figure 2b in the paper
+if plot_results:
+    import matplotlib
+    matplotlib.rc('text', usetex=True)
+    matplotlib.rcParams['text.latex.preamble'] = [r"\usepackage{amsmath}"]
+    import matplotlib.pyplot as plt
+
+    timing_mean = timing_results.mean(axis=3)
+
+    # Only include those results with runtimes lower than 1000 seconds in average
+    Ade4 = timing_mean[0, 0, timing_mean[0, 0] < 1000]
+    UNIPALS = timing_mean[1, 0, timing_mean[1, 0] < 1000]
+    KERNEL = timing_mean[2, 0, timing_mean[2, 0] < 1000]
+    NIPALS = timing_mean[3, 0, timing_mean[3, 0] < 1000]
+    SIMPLS = timing_mean[4, 0, timing_mean[4, 0] < 1000]
+
+    plt.figure(figsize=(6, 4))
+    plt.plot(Ade4, 'b', label='Ade4')
+    plt.plot(UNIPALS, 'c', label='UNIPALS')
+    plt.plot(KERNEL, 'g', label='KERNEL')
+    plt.plot(NIPALS, 'r', label='NIPALS')
+    plt.plot(SIMPLS, 'm', label='SIMPLS\n(Non-Multiblock)')
+    plt.legend(fontsize=13)
+    plt.xlabel(r'Total number of samples $N$ and variables $P$ in $\boldsymbol{X}$', fontsize=14)
+    plt.ylabel('Time in seconds', fontsize=14)
+    xticks = np.arange(-1, 200, 50)
+    xticks[0] = 0
+    plt.xticks(xticks, xticks * 100 + 100, fontsize=13)
+    plt.yticks(fontsize=13)
+    plt.ylim(0, 400)
+    plt.xlim(0, 200)
+    plt.tight_layout()
+    plt.show()
