### 2019-02-06 #1190 @lvermue
@brainstorm @arokem I just updated the runtime analysis scripts. They now contain the code that was used to create the plots shown in the paper. See
https://github.com/DTUComputeStatisticsAndDataAnalysis/MBPLS/commit/abc9bf5771216a2a89bf4354d5f7f9f395d755f4

Furthermore, I had a look at the missing/invalid DOIs.
**Missing DOI**
I could not find any DOI for the paper "Scikit-learn: Machine learning in Python" and the suggested one is definitely wrong. Is there another way to resolve this issue?
**Invalid DOI**
I have triple checked this DOI and cannot see what should be wrong with it. Any idea on how to solve fix one? :)
