### 2019-02-07 #1190 @brainstorm
@lvermue I suspect it has to do with "special characters" in the URL, @arfon, did you encounter those errors in the whedon DOI checker parser before?
