### 2019-02-07 #1190 @arfon
OK, @whedon now recognizes `http://doi.org/10.1002/(SICI)1099-128X(199809/10)12:5<301::AID-CEM515>3.0.CO;2-S` as a valid DOI.
