### 2019-02-08 4fc3e51
Add new version number to conclude JOSS review changes
The new version number is necessary to allow to push the new code to the
PIPY respository with identical version numbers.
	modified:   mbpls/__init__.py
