### 2019-02-08 #1190 @lvermue
@brainstorm @arokem Thank you for the swift and constructive review process!
@arfon Thanks for the super fast fix!

Here the zenodo DOI:
https://doi.org/10.5281/zenodo.2560303
